[rebase] moving uds to uds_blocking for rebase
